Safari-611.1.10/LayoutTests/crypto/subtle/ec-import-jwk-key-export-pkcs8-key-expected.txt

Test importing a jwk private ECDH key and then export it in PKCS8 format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS bytesToHexString(privateKey) is expectedPkcs8KeyHex
PASS successfullyParsed is true

TEST COMPLETE
